what the sigma theres still 1 issue

+++ Dynamidrive/Models/ZStackData.swift
@@ -8,7 +8,7 @@ struct ZStackData: Identifiable {
     var isPlaying: Bool = false
     var showingFilePicker: Bool = false
     var offset: CGFloat = 0
-    var volume: Double = 0.0
+    var volume: Float = 0.0
     
     init(id: Int) {
         self.id = id
